Fix a function prototype broken on 32-bit machines
This was a simple error: I just interchanged uint64_t and size_t on
either half of a function prototype.

+++ src/libmhng/daemon/message.c++
@@ -53,7 +53,7 @@ daemon::message_ptr daemon::message::sync(int32_t seconds_since)
     return out;
 }
 
-daemon::message_ptr daemon::message::idle(uint64_t uid)
+daemon::message_ptr daemon::message::idle(size_t uid)
 {
     auto out = std::make_shared<daemon::message>(daemon::message_type::IDLE);
     out->wire.idle.uid = uid;
